Move ElementPositionTracker out of buttons.
