+++ examples/multiple_windows/lib/app/tooltip_button.dart
@@ -7,84 +7,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/_window.dart';
+import 'element_position_tracker.dart';
 import 'models.dart';
-
-/// Manages all [_ElementPositionTracker] instances.
-///
-/// This class is a singleton because it needs to hook into the global frame
-/// callbacks to update all registered trackers. This callback may only be
-/// registered once, so we use a singleton to ensure that.
-class _ElementPositionTrackerManager {
-  _ElementPositionTrackerManager._() {
-    WidgetsBinding.instance.addPersistentFrameCallback((_) {
-      final trackersCopy = List<_ElementPositionTracker>.from(
-        _trackers,
-        growable: false,
-      );
-      for (final tracker in trackersCopy) {
-        tracker._updateSelf();
-      }
-    });
-  }
-
-  static final _instance = _ElementPositionTrackerManager._();
-  static _ElementPositionTrackerManager get instance => _instance;
-  final List<_ElementPositionTracker> _trackers = <_ElementPositionTracker>[];
-
-  void add(_ElementPositionTracker tracker) {
-    _trackers.add(tracker);
-  }
-
-  void remove(_ElementPositionTracker tracker) {
-    _trackers.remove(tracker);
-  }
-}
-
-/// Tracks global rect of an [Element].
-class _ElementPositionTracker {
-  _ElementPositionTracker({required this.element});
-
-  final BuildContext element;
-  Rect? _lastReportedRect;
-
-  /// Returns current global rect for tracked element or `null` if not available.
-  Rect? getGlobalRect() {
-    final rect = _getGlobalRect();
-    _lastReportedRect = rect;
-    return rect;
-  }
-
-  /// Callback invoked every time the global position of the tracked element changes
-  /// compared to last result of [getGlobalRect].
-  void Function(Rect rect)? onGlobalRectChange;
-
-  Rect? _getGlobalRect() {
-    if (!element.mounted) {
-      return null;
-    }
-    final renderBox = element.findRenderObject();
-    if (renderBox is! RenderBox) {
-      return null;
-    }
-
-    final transform = renderBox.getTransformTo(null);
-    final rect = Offset.zero & renderBox.size;
-    final globalRect = MatrixUtils.transformRect(transform, rect);
-    return globalRect;
-  }
-
-  void _updateSelf() {
-    final rect = _getGlobalRect();
-    if (rect == null) {
-      _ElementPositionTrackerManager.instance.remove(this);
-      return;
-    }
-    if (_lastReportedRect != rect) {
-      _lastReportedRect = rect;
-    }
-    onGlobalRectChange?.call(rect);
-  }
-}
 
 class TooltipButton extends StatefulWidget {
   const TooltipButton({super.key, required this.parentController});
@@ -97,14 +21,12 @@ class TooltipButton extends StatefulWidget {
 
 class _TooltipButtonState extends State<TooltipButton> {
   TooltipWindowController? _tooltipController;
-  _ElementPositionTracker? _tooltipTracker;
+  ElementPositionTracker? _tooltipTracker;
   final GlobalKey _tooltipButtonKey = GlobalKey();
 
   @override
   void dispose() {
-    if (_tooltipTracker != null) {
-      _ElementPositionTrackerManager.instance.remove(_tooltipTracker!);
-    }
+    _tooltipTracker?.dispose();
     super.dispose();
   }
 
@@ -115,19 +37,16 @@ class _TooltipButtonState extends State<TooltipButton> {
     // Toggle tooltip visibility.
     if (_tooltipController != null) {
       _tooltipController!.destroy();
-      if (_tooltipTracker != null) {
-        _ElementPositionTrackerManager.instance.remove(_tooltipTracker!);
-      }
+      _tooltipTracker?.dispose;
       setState(() {
         _tooltipController = null;
         _tooltipTracker = null;
       });
     } else {
       // Tooltip is not shown, show it.
-      final tracker = _ElementPositionTracker(
+      final tracker = ElementPositionTracker(
         element: _tooltipButtonKey.currentContext!,
       );
-      _ElementPositionTrackerManager.instance.add(tracker);
       final UniqueKey key = UniqueKey();
       final controller = TooltipWindowController(
         anchorRect: tracker.getGlobalRect()!,
@@ -135,7 +54,7 @@ class _TooltipButtonState extends State<TooltipButton> {
         delegate: _TooltipWindowControllerDelegate(
           onDestroyed: () {
             windowManager.remove(key);
-            _ElementPositionTrackerManager.instance.remove(tracker);
+            tracker.dispose();
             if (mounted) {
               setState(() {
                 _tooltipController = null;
